--- src/Source/SourceInterface.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace webignition\CssValidatorWrapper\Source;
-
-interface SourceInterface
-{
-    public function getUri(): string;
-    public function isAvailable(): bool;
-}
