--- src/Source/UnavailableSource.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace webignition\CssValidatorWrapper\Source;
-
-class UnavailableSource extends AbstractSource
-{
-    public function isAvailable(): bool
-    {
-        return false;
-    }
-}
